[BE] refactor: request validation 추가 #338
