+++ backend/src/main/java/com/onair/hearit/auth/dto/request/LoginRequest.java
@@ -1,7 +1,12 @@
 package com.onair.hearit.auth.dto.request;
 
+import jakarta.validation.constraints.NotBlank;
+
 public record LoginRequest(
+        @NotBlank
         String localId,
+
+        @NotBlank
         String password
 ) {
 }
