+++ backend/src/main/java/com/onair/hearit/auth/dto/request/TokenReissueRequest.java
@@ -1,6 +1,9 @@
 package com.onair.hearit.auth.dto.request;
 
+import jakarta.validation.constraints.NotBlank;
+
 public record TokenReissueRequest(
+        @NotBlank
         String refreshToken
 ) {
 }
