+++ backend/src/main/java/com/onair/hearit/auth/dto/response/OAuthUserInfoResponse.java
@@ -1,8 +1,8 @@
-package com.onair.hearit.auth.dto.request;
+package com.onair.hearit.auth.dto.response;
 
 import com.onair.hearit.auth.domain.OAuthProvider;
 
-public record OAuthUserInfo(
+public record OAuthUserInfoResponse(
         String id,
         String nickname,
         String profileImageUrl,
