+++ backend/src/main/java/com/onair/hearit/auth/infrastructure/oauth/kakao/KakaoOAuthService.java
@@ -2,8 +2,8 @@ package com.onair.hearit.auth.infrastructure.oauth.kakao;
 
 import com.onair.hearit.auth.application.OAuthService;
 import com.onair.hearit.auth.domain.OAuthProvider;
-import com.onair.hearit.auth.dto.request.OAuthUserInfo;
 import com.onair.hearit.auth.dto.response.KakaoUserInfoResponse;
+import com.onair.hearit.auth.dto.response.OAuthUserInfoResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestClient;
@@ -20,9 +20,9 @@ public class KakaoOAuthService implements OAuthService {
     }
 
     @Override
-    public OAuthUserInfo fetchUser(String accessToken) {
+    public OAuthUserInfoResponse fetchUser(String accessToken) {
         KakaoUserInfoResponse response = getUserInfo(accessToken);
-        return new OAuthUserInfo(response.id(), response.nickname(), response.profileImage(), provider());
+        return new OAuthUserInfoResponse(response.id(), response.nickname(), response.profileImage(), provider());
     }
 
     private KakaoUserInfoResponse getUserInfo(String kakaoAccessToken) {
